backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
aiofiles
python-dotenv
gunicorn